requirements.txt
Django==2.0.4
gunicorn==19.6.0
pytz==2018.4
whitenoise==3.2.1
dj-database-url==0.5.0
psycopg2==2.7.4
djangorestframework==3.8.2